test(entityManager): split component and lifecycle suites
